Extract shared createMatch helper into SeedsMatches trait
createMatch was duplicated across ApiTest and SeasonApiTest. Move the
array-based version into a Tests\Concerns\SeedsMatches trait and have both
test classes use it; the array form covers the single-player case via
[$player]. Drops now-unused imports from ApiTest.
